pyproject.toml
[tool.mypy]
# TODO: widen range of files as we fix issues
namespace_packages = true

[[tool.mypy.overrides]]
module = ['ruamel', 'networkx']
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 99
target-version = ['py38']

[tool.ruff]
# Allow lines to be as long as 120 characters.
line-length = 120

[tool.poetry]
name = "dbt-meshify"
version = "0.3.0"
description = "a package to upgrade dbt packages to the dbt mesh framework"
authors = [
    "Dave Connors <dave.connors@fishtownanalytics.com>",
    "Grace Goheen <grace.goheen@dbtlabs.com>",
]
license = "Apache 2.0"
readme = "README.md"
packages = [{ include = "dbt_meshify" }]

[tool.poetry.dependencies]
python = "^3.9"
dbt-core = "^1.5.0"
click = "^8.1.3"
dbt-postgres = { version = "^1.5.0", optional = true }
ruamel-yaml = "^0.17.31"
loguru = "^0.7.0"

[tool.poetry.extras]
postgres = ['dbt-postgres']

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
dbt-duckdb = "^1.5.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mike = "^1.1.2"
mkdocs = "^1.4.3"
mkdocs-click = "^0.8.0"
mkdocs-material = "^9.1.12"
mypy = "^1.3.0"
pre-commit = "^3.3.1"
pytest = "^7.3.1"
types-pyyaml = "^6.0.12.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dbt-meshify = "dbt_meshify.main:cli"
